LoongArch: Do not define legacy built-in macros for LA32
These all-caps built-in macros, while defined in the 2022.01.13 version
of the original *LoongArch Toolchain Conventions* document [1], are
deprecated since version 2023.04.20 [2] by [3], and since the recent
*Software Development and Build Convention for LoongArch Architectures*
document was released in July 2023, they have completely disappeared.
Now that it is nearing the end of 2025 and a new LoongArch target is
added, better not include the cruft for not deviating from the official
specs, and a cleaner ecosystem.

[1]: https://github.com/loongson/LoongArch-Documentation/blob/2022.01.13/docs/LoongArch-toolchain-conventions-EN.adoc#cc-preprocessor-definitions
[2]: https://github.com/loongson/LoongArch-Documentation/blob/2023.04.20/docs/LoongArch-toolchain-conventions-EN.adoc#cc-preprocessor-built-in-macro-definitions
[3]: https://github.com/loongson/LoongArch-Documentation/pull/28
[4]: https://github.com/loongson/la-softdev-convention/blob/v0.1/la-softdev-convention.adoc#83-architecture-related-cc-pre-defined-macros

gcc/ChangeLog:

	* config/loongarch/loongarch-c.cc
	(loongarch_define_unconditional_macros): Do not define legacy
	compatibility macros for LA32.
	(loongarch_update_cpp_builtins): Ditto.

Signed-off-by: WANG Xuerui <git@xen0n.name>